ref(onboarding): Move test helper to sentry-test
